Macro rtt_target::rprintln

macro_rules! rprintln {
    (=> $terminal:expr) => { ... };
    (=> $terminal:expr, $fmt:expr) => { ... };
    (=> $terminal:expr, $fmt:expr, $($arg:tt)*) => { ... };
    () => { ... };
    ($fmt:expr) => { ... };
    ($fmt:expr, $($arg:tt)*) => { ... };
}

Prints to the print RTT channel, with a newline. Works just like the standard println.

Before use the print channel has to be set with rtt_init_print or set_print_channel. If the channel isn't set, the message is silently discarded.

The macro also supports output to multiple virtual terminals on the channel. Use the syntax rprintln!(=> 1, "Hello!"); to write to terminal number 1, for example. Terminal numbers range from 0 to 15.