Macro rtt_target::rtt_init_print

macro_rules! rtt_init_print {
    ($mode : ident, $size : literal) => { ... };
    ($mode : ident) => { ... };
    () => { ... };
}

Initializes RTT with a single up channel and sets it as the print channel for the printing macros.

The optional arguments specify the blocking mode (default: NoBlockSkip) and size of the buffer in bytes (default: 1024). See rtt_init for more details.

This macro is defined only if the set_print_channel function is available, i.e. if you have enabled a platform support feature.