Type Definition rtt_target::CriticalSectionFunc

type CriticalSectionFunc = fn(arg: *mut (), f: fn(arg: *mut ()));

Type-erased critical section function used to synchronize printing.

When called, the function must establish a critical section and call f within it, passing arg as the argument.